Android fixes (#222)
* Taking out uiHandler posts of parseAST nad SyncState, this will reduce the amount of work on the uiHandler'

* Minor cleanup

* Adding remove listener feature to Event emitter

* Removing hardware caching and getting floats for y and x

* Removing component did mount calls for functional components

* Run Pretty

* Cleanup

* Removing unused imports;

* Removing unNecessary logs

* Running pretty on JS files

+++ android/syrnative/src/main/java/syr/js/org/syrnative/SyrScrollview.java
@@ -7,6 +7,8 @@ import android.widget.ScrollView;
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import java.math.BigDecimal;
+
 /**
  * Created by dereanderson on 1/10/18.
  */
@@ -17,7 +19,7 @@ public class SyrScrollview implements SyrBaseModule, SyrComponent {
     public View render(JSONObject component, Context context, View instance) {
 
         ScrollView scrollview;
-        if(instance != null) {
+        if (instance != null) {
             scrollview = (ScrollView) instance;
         } else {
             scrollview = new ScrollView(context);
@@ -54,11 +56,11 @@ public class SyrScrollview implements SyrBaseModule, SyrComponent {
                 SyrStyler.styleView(scrollview, style);
 
                 if (style.has("left")) {
-                    scrollview.setX(style.getInt("left"));
+                    scrollview.setX(BigDecimal.valueOf(style.getDouble("left")).floatValue());
                 }
 
                 if (style.has("top")) {
-                    scrollview.setY(style.getInt("top"));
+                    scrollview.setY(BigDecimal.valueOf(style.getDouble("top")).floatValue());
                 }
 
             }
